Call reserveDebt for all clearinghouse versions
