+++ .subgraph-version
@@ -1,1 +1,1 @@
-1.4.0
+1.4.1
