+++ src/constants.ts
@@ -1,0 +1,2 @@
+export const COOLER_LOANS_CLEARINGHOUSE_V1 = "0xD6A6E8d9e82534bD65821142fcCd91ec9cF31880".toLowerCase();
+export const COOLER_LOANS_CLEARINGHOUSE_V1_1 = "0xE6343ad0675C9b8D3f32679ae6aDbA0766A2ab4c".toLowerCase();
